chore: rename Index type files including Index word
