+++ src/utils/data/data.ts
@@ -1,7 +1,7 @@
 import urlJoin from "url-join"
 import { LINKS } from "../constants"
-import { IndexBySchoolYear } from "../types/data/parsed/Index/BySchoolYear"
-import { IndexBySchoolYearCourse } from "../types/data/parsed/Index/BySchoolYearCourse"
+import { IndexBySchoolYear } from "../types/data/parsed/Index/IndexBySchoolYear"
+import { IndexBySchoolYearCourse } from "../types/data/parsed/Index/IndexBySchoolYearCourse"
 import School from "../types/data/School"
 import Ranking from "../types/data/parsed/Ranking"
 import JsonParser from "./jsonParser"
