+++ src/utils/types/data/parsed/Index/IndexBySchoolYear.ts
