+++ src/utils/types/data/parsed/Index/IndexBySchoolYearCourse.ts
