+++ src/utils/types/data/parsed/Index/IndexByYearSchool.ts
